Increase task completion timeout from 0.5 => 1.0 sec
As seen at https://travis-ci.org/ReactiveX/RxJavaFX/builds/41017012,
`JavaFxSchedulerTest.testPeriodicScheduling` still fails under Travis
CI, though it is assumed not because of negative delay time values but
because of actual failure to execute tasks in the given half-second
period. This failure is unreproducible on a 4-core MBP, so the 1 second
timeout is a guess.